Split web API into feature mixins
api.py mixed HTTP routing, overview and history building, probing, subscription
maintenance and serialisation in 1537 lines. Move each concern into a mixin and
leave register(), the handle_* adapters and the response helpers (450 lines).

Pure move: NitterWebAPI keeps all 79 members. Five staticmethods referenced the
host class by name and now name the mixin that defines the target.

Also drop the unused _parse_float helper.
