+++ plugin_api/api.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import asyncio
 import inspect
-import math
 from collections.abc import Awaitable, Callable
 from typing import Any
 
@@ -13,70 +12,44 @@ try:
     from ..config import (
         config_get,
         configured_merge_tweet_threshold,
-        media_only_unavailable_reason,
-        parse_config_bool,
-        resolve_send_image_attachments,
-        resolve_send_video_attachments,
     )
     from ..config.subscriptions import (
-        ensure_default_import_group,
-        normalize_import_username,
         save_subscription_config,
-        set_import_group_users,
-        sync_import_config_groups,
     )
     from ..scheduler import ScheduleGroup
-    from ..storage import (
-        PushHistoryGroupSummary,
-        PushHistoryRecord,
-    )
-    from ..delivery import PlatformResolver, parse_umo
-    from ..shared import TweetItem, normalize_username
-    from ..shared.group_ids import normalize_stable_group_id
-    from ..media_support.html_backend.query import (
-        MAX_QUERY_LENGTH,
-        normalize_query,
-        query_kind,
-    )
-    from ..media_support.network import UnsafeUrlError, validate_http_url
+    from .api_history import WebAPIHistoryMixin
+    from .api_overview import WebAPIOverviewMixin
+    from .api_probe import WebAPIProbeMixin
+    from .api_serializers import WebAPISerializersMixin
+    from .api_subscriptions import WebAPISubscriptionsMixin
     from .groups import WebUIGroupEditor
 except ImportError:
     from config import (
         config_get,
         configured_merge_tweet_threshold,
-        media_only_unavailable_reason,
-        parse_config_bool,
-        resolve_send_image_attachments,
-        resolve_send_video_attachments,
     )
     from config.subscriptions import (
-        ensure_default_import_group,
-        normalize_import_username,
         save_subscription_config,
-        set_import_group_users,
-        sync_import_config_groups,
     )
-    from delivery import PlatformResolver, parse_umo
     from scheduler import ScheduleGroup
-    from storage import (
-        PushHistoryGroupSummary,
-        PushHistoryRecord,
-    )
-    from shared import TweetItem, normalize_username
-    from shared.group_ids import normalize_stable_group_id
-    from media_support.html_backend.query import (
-        MAX_QUERY_LENGTH,
-        normalize_query,
-        query_kind,
-    )
-    from media_support.network import UnsafeUrlError, validate_http_url
+    from plugin_api.api_history import WebAPIHistoryMixin
+    from plugin_api.api_overview import WebAPIOverviewMixin
+    from plugin_api.api_probe import WebAPIProbeMixin
+    from plugin_api.api_serializers import WebAPISerializersMixin
+    from plugin_api.api_subscriptions import WebAPISubscriptionsMixin
     from plugin_api.groups import WebUIGroupEditor
 
 
 PLUGIN_NAME = "astrbot_plugin_nitter_tweets"
 
 
-class NitterWebAPI:
+class NitterWebAPI(
+    WebAPIOverviewMixin,
+    WebAPIHistoryMixin,
+    WebAPIProbeMixin,
+    WebAPISubscriptionsMixin,
+    WebAPISerializersMixin,
+):
     """Backend API provider for the AstrBot Plugin Pages dashboard."""
 
     def __init__(self, plugin: Any) -> None:
@@ -216,59 +189,6 @@ class NitterWebAPI:
 
         return await self._json_response(action)
 
-    async def build_overview(self) -> dict[str, Any]:
-        groups = self._schedule_groups()
-
-        total_raw_users = sum(group.users_info.raw_count for group in groups)
-        total_duplicates = sum(len(group.users_info.duplicates) for group in groups)
-        total_invalid_users = sum(
-            len(group.users_info.invalid_entries) for group in groups
-        )
-        counts = {
-            "groups": len(groups),
-            "enabled_groups": sum(1 for group in groups if group.enabled),
-            "watch_users": sum(len(group.users) for group in groups),
-            "raw_watch_users": total_raw_users,
-            "duplicate_watch_users": total_duplicates,
-            "invalid_watch_users": total_invalid_users,
-            "push_targets": sum(len(group.targets) for group in groups),
-            "invalid_push_targets": sum(len(group.invalid_targets) for group in groups),
-        }
-        scheduler_state = {
-            "running": bool(getattr(self.scheduler, "is_running", False)),
-            "schedule_enabled": bool(
-                getattr(self.scheduler, "schedule_enabled", False)
-            ),
-        }
-        features = {
-            "images": resolve_send_image_attachments(self.config),
-            "videos": resolve_send_video_attachments(self.config),
-            "translation": parse_config_bool(
-                config_get(self.config, "translate_enabled", False), False
-            ),
-        }
-        instance_lists = self._configured_instance_lists()
-        instances = list(instance_lists["rss"])
-        return self._ok(
-            scheduler=scheduler_state,
-            counts=counts,
-            features=features,
-            config_summary=self._config_summary(instances, groups, instance_lists),
-            instances=instances,
-            instance_lists=instance_lists,
-            attention_items=self._overview_attention_items(
-                counts, scheduler_state, instances, groups
-            ),
-            terminology=self._terminology(),
-        )
-
-    async def build_groups(self) -> dict[str, Any]:
-        groups = self._schedule_groups()
-        return self._ok(
-            groups=[self._serialize_group(group) for group in groups],
-            terminology=self._terminology(),
-        )
-
     async def create_group(self, data: dict[str, Any]) -> dict[str, Any]:
         result = self._group_editor().create_group(data)
         if not result.get("success"):
@@ -338,190 +258,6 @@ class NitterWebAPI:
                 payload["message"] = f"分组已删除，但数据库同步失败：{sync_error}"
         return payload
 
-    async def probe_targets(self, data: dict[str, Any]) -> dict[str, Any]:
-        group_id = self._data_text(data, "group_id")
-        group: ScheduleGroup | None = None
-        if group_id:
-            group, error = self._resolve_group(group_id)
-            if error:
-                return self._error(error)
-
-        targets = self._data_text_list(data, "target_umos")
-        if not targets and group is not None:
-            targets = [*group.targets, *group.invalid_targets]
-        if not targets:
-            return self._error("请填写要检测的推送目标")
-
-        results = [self._probe_target(target) for target in targets]
-        valid_count = sum(1 for item in results if item["valid"])
-        invalid_count = len(results) - valid_count
-        found_count = sum(1 for item in results if item.get("platform_found"))
-        merge_count = sum(1 for item in results if item.get("supports_merged_forward"))
-        return self._ok(
-            group_id=group.group_id if group is not None else group_id,
-            group_name=group.name if group is not None else "",
-            targets=results,
-            summary={
-                "total": len(results),
-                "valid": valid_count,
-                "invalid": invalid_count,
-                "platform_found": found_count,
-                "supports_merged_forward": merge_count,
-            },
-        )
-
-    def _probe_target(self, target_umo: str) -> dict[str, Any]:
-        target_umo = str(target_umo or "").strip()
-        platform_id, message_type, session_id = parse_umo(target_umo)
-        if not platform_id or not message_type or not session_id:
-            return {
-                "umo": target_umo,
-                "valid": False,
-                "error": "推送目标必须是 /sid 返回的完整 UMO：platform:MessageType:session_id",
-                "platform_id": platform_id,
-                "message_type": message_type,
-                "session_id": session_id,
-                "platform_kind": "unknown",
-                "platform_found": False,
-                "supports_merged_forward": False,
-            }
-
-        context = getattr(self.scheduler, "context", None) or getattr(
-            self.plugin, "context", None
-        )
-        profile = PlatformResolver().from_umo(context, target_umo)
-        supports_merged_forward = False
-        sender = getattr(self.plugin, "sender", None) or getattr(
-            self.scheduler, "sender", None
-        )
-        supports = getattr(sender, "supports_merged_forward_for_umo", None)
-        if callable(supports):
-            try:
-                supports_merged_forward = bool(supports(context, target_umo))
-            except Exception:
-                supports_merged_forward = False
-
-        return {
-            "umo": target_umo,
-            "valid": True,
-            "error": "",
-            "platform_id": platform_id,
-            "message_type": message_type,
-            "session_id": session_id,
-            "platform_kind": self._platform_kind(profile),
-            "platform_types": list(profile.platform_types),
-            "platform_found": profile.platform is not None,
-            "supports_merged_forward": supports_merged_forward,
-        }
-
-    @staticmethod
-    def _platform_kind(profile: Any) -> str:
-        if getattr(profile, "is_lark", False):
-            return "lark"
-        if getattr(profile, "is_telegram", False):
-            return "telegram"
-        if getattr(profile, "is_onebot", False):
-            return "onebot"
-        if getattr(profile, "is_known_non_onebot", False):
-            return "non_onebot"
-        return "default"
-
-    async def build_history(
-        self,
-        group_id: str = "",
-        username: str = "",
-        limit: int = 10,
-        offset: int = 0,
-    ) -> dict[str, Any]:
-        group_id = str(group_id or "").strip()
-        username = str(username or "").strip().lstrip("@")
-        limit = max(1, min(int(limit or 10), 50))
-        offset = max(0, int(offset or 0))
-        if group_id:
-            group, error = self._resolve_group(group_id)
-            if error:
-                return self._error(error)
-            group_id = group.group_id
-        total_count = await self.storage.count_push_history(group_id, username)
-        total_pages = max(1, math.ceil(total_count / limit))
-        records = await self.storage.get_push_history(
-            group_id,
-            username,
-            limit + 1,
-            offset,
-        )
-        group_names = {group.group_id: group.name for group in self._schedule_groups()}
-        groups_by_id = {group.group_id: group for group in self._schedule_groups()}
-        grouped_records = self._group_history_records(
-            records, group_names, groups_by_id
-        )
-        has_next = len(grouped_records) > limit
-        visible_records = grouped_records[:limit]
-        return self._ok(
-            selected_group_id=group_id,
-            selected_username=username,
-            limit=limit,
-            offset=offset,
-            page=(offset // limit) + 1,
-            total_count=total_count,
-            total_pages=total_pages,
-            has_prev=offset > 0,
-            has_next=has_next,
-            prev_offset=max(0, offset - limit),
-            next_offset=offset + limit if has_next else offset,
-            records=visible_records,
-            terminology=self._terminology(),
-        )
-
-    async def build_history_orphans(self) -> dict[str, Any]:
-        groups = self._schedule_groups()
-        configured_ids = {normalize_stable_group_id(group.group_id) for group in groups}
-        summaries = await self.storage.get_push_history_group_summaries()
-        orphans = [
-            self._serialize_history_group_summary(summary)
-            for summary in summaries
-            if normalize_stable_group_id(summary.group_id) not in configured_ids
-        ]
-        return self._ok(
-            orphans=orphans,
-            terminology=self._terminology(),
-        )
-
-    async def delete_history_orphan(self, data: dict[str, Any]) -> dict[str, Any]:
-        raw_group_id = self._data_text(data, "group_id")
-        if not raw_group_id:
-            return self._error("请选择要清理的分组 ID")
-        group_id = normalize_stable_group_id(raw_group_id)
-        configured_ids = {
-            normalize_stable_group_id(group.group_id)
-            for group in self._schedule_groups()
-        }
-        if group_id in configured_ids:
-            return self._error(f"分组仍存在，不能作为失效分组清理：{group_id}")
-        if self._data_text(data, "confirm") != "DELETE":
-            return self._error("清理失效分组运行数据需要显式确认")
-
-        summary = await self.storage.delete_orphan_group_runtime_data(group_id)
-        return self._ok(
-            group_id=group_id,
-            summary=summary,
-        )
-
-    async def replay_history(self, data: dict[str, Any]) -> dict[str, Any]:
-        record_id = self._parse_int(
-            self._data_text(data, "record_id") or data.get("record_id"),
-            0,
-            minimum=0,
-            maximum=10_000_000_000,
-        )
-        if record_id <= 0:
-            return self._error("请选择要重新推送的记录")
-        target_umos = self._data_text_list(data, "target_umos")
-        result = await self.scheduler.replay_push_history(record_id, target_umos)
-        if not result.get("success"):
-            return result
-        return self._ok(**result)
-
     async def run_check(self, data: dict[str, Any]) -> dict[str, Any]:
         group_id = self._data_text(data, "group_id") or self._data_text(
             data, "group_name"
@@ -569,270 +305,6 @@ class NitterWebAPI:
             legacy_deleted=bool(legacy_deleted),
             warning=("推送记录已清理；关注账号、推送目标和媒体文件不会被删除。"),
         )
-
-    async def import_subscriptions(self, data: dict[str, Any]) -> dict[str, Any]:
-        entries = self._subscription_entries(data)
-        if not entries:
-            return self._error("请填写要导入的关注账号")
-
-        group, error = self._subscription_group(data)
-        if error:
-            return self._error(error)
-
-        existing_users = list(group.users)
-        seen = {user.lower() for user in existing_users}
-        added: list[str] = []
-        duplicates: list[str] = []
-        invalid: list[str] = []
-
-        for raw in entries:
-            username = normalize_import_username(raw)
-            if not username:
-                invalid.append(raw)
-                continue
-            key = username.lower()
-            if key in seen:
-                duplicates.append(raw)
-                continue
-            seen.add(key)
-            added.append(username)
-
-        if added:
-            set_import_group_users(
-                self.config,
-                self.scheduler.config_reader,
-                group,
-                [*existing_users, *added],
-            )
-            save_error = self._save_config()
-            sync_error = await sync_import_config_groups(self.scheduler)
-        else:
-            save_error = ""
-            sync_error = ""
-
-        return self._ok(
-            message=self._subscription_message(
-                "导入", bool(added), save_error, sync_error
-            ),
-            summary={
-                "group": self._group_payload_label(group),
-                "input_count": len(entries),
-                "added": added,
-                "duplicates": duplicates,
-                "invalid": invalid,
-                "total_after": len(existing_users) + len(added),
-                "saved": bool(added),
-                "save_error": save_error,
-                "sync_error": sync_error,
-            },
-        )
-
-    async def delete_subscriptions(self, data: dict[str, Any]) -> dict[str, Any]:
-        entries = self._subscription_entries(data)
-        if not entries:
-            return self._error("请填写要删除的关注账号")
-        if len(entries) > 50:
-            return self._error(
-                f"单次最多删除 50 个关注账号，本次输入 {len(entries)} 个"
-            )
-
-        group, error = self._subscription_group(data)
-        if error:
-            return self._error(error)
-
-        existing_users = list(group.users)
-        existing_by_key = {user.lower(): user for user in existing_users}
-        delete_keys: set[str] = set()
-        requested: list[str] = []
-        duplicates: list[str] = []
-        invalid: list[str] = []
-
-        for raw in entries:
-            username = normalize_import_username(raw)
-            if not username:
-                invalid.append(raw)
-                continue
-            key = username.lower()
-            if key in delete_keys:
-                duplicates.append(raw)
-                continue
-            delete_keys.add(key)
-            requested.append(username)
-
-        removed = [
-            existing_by_key[user.lower()]
-            for user in requested
-            if user.lower() in existing_by_key
-        ]
-        missing = [user for user in requested if user.lower() not in existing_by_key]
-        remaining = [user for user in existing_users if user.lower() not in delete_keys]
-
-        if removed:
-            set_import_group_users(
-                self.config,
-                self.scheduler.config_reader,
-                group,
-                remaining,
-            )
-            save_error = self._save_config()
-            sync_error = await sync_import_config_groups(self.scheduler)
-        else:
-            save_error = ""
-            sync_error = ""
-
-        return self._ok(
-            message=self._subscription_message(
-                "删除", bool(removed), save_error, sync_error
-            ),
-            summary={
-                "group": self._group_payload_label(group),
-                "input_count": len(entries),
-                "removed": removed,
-                "missing": missing,
-                "duplicates": duplicates,
-                "invalid": invalid,
-                "total_after": len(remaining),
-                "saved": bool(removed),
-                "save_error": save_error,
-                "sync_error": sync_error,
-            },
-        )
-
-    async def probe_mirror(self, data: dict[str, Any]) -> dict[str, Any]:
-        instance = self._data_text(data, "instance")
-        try:
-            # Keep the UI probe deterministic even when DNS is temporarily
-            # unavailable; the actual opener repeats strict DNS and redirect
-            # validation immediately before connecting.
-            instance = validate_http_url(instance, resolve_dns=False).rstrip("/")
-        except UnsafeUrlError:
-            return self._error(
-                "请填写完整 Nitter 镜像站地址，例如 https://nitter.net"
-            )
-
-        mode = self._data_text(data, "mode") or "blogger_rss"
-        mode = mode.strip().lower().replace("-", "_")
-        if mode not in {"blogger_rss", "search"}:
-            if mode == "blogger_html":
-                return self._error(
-                    "博主 HTML 回退已移除；请用 blogger_rss 或 search"
-                )
-            return self._error("mode 仅支持 blogger_rss / search")
-
-        limit = self._parse_int(
-            data.get("limit"),
-            int(getattr(self.plugin, "default_limit", 5) or 5),
-            minimum=1,
-            maximum=50,
-        )
-
-        try:
-            if mode == "blogger_rss":
-                username = normalize_username(
-                    self._data_text(data, "username")
-                    or self._data_text(data, "query")
-                    or "nasa"
-                )
-                if not username:
-                    return self._error("关注账号格式无效")
-                used_instance, tweets = (
-                    await self.plugin.nitter.fetch_tweets_from_instance(
-                        instance,
-                        username,
-                        limit,
-                    )
-                )
-                subject = username
-                kind = ""
-            else:
-                raw_query = (
-                    self._data_text(data, "query")
-                    or self._data_text(data, "username")
-                    or ""
-                ).strip()
-                if len(raw_query) > MAX_QUERY_LENGTH:
-                    return self._error(
-                        f"搜索内容过长（最多 {MAX_QUERY_LENGTH} 字符）"
-                    )
-                query = normalize_query(raw_query)
-                if not query:
-                    return self._error("请填写搜索内容（#标签 或短语）")
-                html_backend = getattr(self.plugin, "html_backend", None)
-                if html_backend is None:
-                    return self._error("HTML 后端未初始化")
-                if not bool(getattr(getattr(html_backend, "config", None), "search_enabled", True)):
-                    return self._error("search_enabled 已关闭")
-                kind = query_kind(query)
-                used_instance, tweets = await asyncio.to_thread(
-                    html_backend.search,
-                    query,
-                    limit,
-                    kind=kind,
-                    instance=instance,
-                )
-                subject = query
-                username = query
-        except Exception as exc:
-            logger.warning(
-                "[NitterTweets] WebUI 镜像测试失败: "
-                f"mode={mode}, instance={instance}, error={exc}"
-            )
-            if mode == "search":
-                return self._error(
-                    f"通过 {instance} 搜索失败：实例不可达、被限流，或搜索门禁未通过。"
-                )
-            return self._error(
-                f"通过 {instance} 获取失败：Nitter 暂时不可用，或用户没有公开 RSS。"
-            )
-
-        return self._ok(
-            mode=mode,
-            instance=used_instance,
-            username=username if mode != "search" else "",
-            query=subject if mode == "search" else "",
-            kind=kind,
-            subject=subject,
-            limit=limit,
-            tweet_count=len(tweets),
-            tweets=[self._serialize_probe_tweet(tweet) for tweet in tweets[:limit]],
-        )
-
-    def _configured_instance_lists(self) -> dict[str, list[str]]:
-        """Three config lists for mirror probe UI (deduped, order preserved)."""
-        rss = list(getattr(getattr(self.plugin, "nitter", None), "instances", []) or [])
-        html_backend = getattr(self.plugin, "html_backend", None)
-        search: list[str] = []
-        if html_backend is not None:
-            cfg = getattr(html_backend, "config", None)
-            if cfg is not None:
-                search = list(getattr(cfg, "search_instances", []) or [])
-        # Do not use load_instances() here: empty config must stay empty
-        # (load_instances falls back to DEFAULT_INSTANCES / nitter.net).
-        if not search:
-            search = list(config_get(self.config, "search_instances", []) or [])
-        return {
-            "rss": self._dedupe_instances(rss),
-            # Kept for dashboard API shape; blogger HTML pool is always empty.
-            "blogger_html": [],
-            "search": self._dedupe_instances(search),
-        }
-
-    @staticmethod
-    def _dedupe_instances(values: list[str]) -> list[str]:
-        seen: set[str] = set()
-        out: list[str] = []
-        for raw in values:
-            item = str(raw or "").strip().rstrip("/")
-            if not item:
-                continue
-            if not item.startswith(("http://", "https://")):
-                item = f"https://{item}"
-            key = item.rstrip("/").lower()
-            if key in seen:
-                continue
-            seen.add(key)
-            out.append(item.rstrip("/"))
-        return out
 
     @property
     def config(self):
@@ -885,42 +357,8 @@ class NitterWebAPI:
             return None, f"未找到分组：{group_id}"
         return group, ""
 
-    def _subscription_group(
-        self, data: dict[str, Any]
-    ) -> tuple[ScheduleGroup | None, str]:
-        group_id = self._data_text(data, "group_id") or self._data_text(
-            data, "group_name"
-        )
-        if group_id:
-            return self._resolve_group(group_id)
-        return ensure_default_import_group(
-            self.config,
-            self.scheduler.config_reader,
-        ), ""
-
-    def _subscription_entries(self, data: dict[str, Any]) -> list[str]:
-        raw = data.get("entries", data.get("users", ""))
-        if isinstance(raw, list):
-            return [str(item).strip() for item in raw if str(item).strip()]
-        return [item.strip() for item in str(raw or "").split(",") if item.strip()]
-
     def _save_config(self) -> str:
         return save_subscription_config(self.config)
-
-    @staticmethod
-    def _subscription_message(
-        action: str,
-        changed: bool,
-        save_error: str = "",
-        sync_error: str = "",
-    ) -> str:
-        if not changed:
-            return f"{action}完成，配置未改动"
-        if save_error:
-            return f"{action}完成，但配置保存失败：{save_error}"
-        if sync_error:
-            return f"{action}完成，但数据库同步失败：{sync_error}"
-        return f"{action}完成"
 
     def _config_summary(
         self,
@@ -985,531 +423,6 @@ class NitterWebAPI:
         if group is None:
             return fallback
         return getattr(group, name, fallback)
-
-    @staticmethod
-    def _overview_attention_items(
-        counts: dict[str, int],
-        scheduler_state: dict[str, bool],
-        instances: list[str],
-        groups: list[ScheduleGroup] | None = None,
-    ) -> list[dict[str, str]]:
-        items: list[dict[str, str]] = []
-        if not scheduler_state.get("running", False):
-            items.append(
-                {
-                    "key": "scheduler_not_running",
-                    "level": "warning",
-                    "title": "调度器未运行",
-                    "detail": "后台检查不会自动执行。",
-                }
-            )
-        if not scheduler_state.get("schedule_enabled", False):
-            items.append(
-                {
-                    "key": "schedule_disabled",
-                    "level": "warning",
-                    "title": "后台检查总开关关闭",
-                    "detail": "定时检查不会自动触发。",
-                }
-            )
-        if not instances:
-            items.append(
-                {
-                    "key": "no_instances",
-                    "level": "warning",
-                    "title": "未配置 Nitter 实例",
-                    "detail": "手动查询、镜像测试和后台检查都需要可用实例。",
-                }
-            )
-        if int(counts.get("groups", 0)) <= 0:
-            items.append(
-                {
-                    "key": "no_groups",
-                    "level": "info",
-                    "title": "没有用户分组",
-                    "detail": "关注账号和推送目标需要先在 AstrBot 设置页配置。",
-                }
-            )
-        elif int(counts.get("enabled_groups", 0)) <= 0:
-            items.append(
-                {
-                    "key": "no_enabled_groups",
-                    "level": "warning",
-                    "title": "没有启用的用户分组",
-                    "detail": "后台检查不会处理已停用分组。",
-                }
-            )
-        if int(counts.get("invalid_push_targets", 0)) > 0:
-            items.append(
-                {
-                    "key": "invalid_push_targets",
-                    "level": "warning",
-                    "title": "存在无效推送目标",
-                    "detail": f"{counts['invalid_push_targets']} 个推送目标未通过 UMO 校验。",
-                }
-            )
-        if int(counts.get("invalid_watch_users", 0)) > 0:
-            items.append(
-                {
-                    "key": "invalid_watch_users",
-                    "level": "warning",
-                    "title": "存在无效关注账号",
-                    "detail": f"{counts['invalid_watch_users']} 个关注账号格式无效。",
-                }
-            )
-        if groups:
-            items.extend(NitterWebAPI._overview_group_diagnostics(groups))
-
-        if not items:
-            items.append(
-                {
-                    "key": "ok",
-                    "level": "ok",
-                    "title": "当前没有需要处理的提示",
-                    "detail": "关键状态未发现异常。",
-                }
-            )
-        return items
-
-    @staticmethod
-    def _overview_group_diagnostics(
-        groups: list[ScheduleGroup],
-    ) -> list[dict[str, str]]:
-        enabled_groups = [group for group in groups if group.enabled]
-        if not enabled_groups:
-            return []
-
-        items: list[dict[str, str]] = []
-        no_watch_users = [
-            group
-            for group in enabled_groups
-            if group.is_blogger_group and not group.users
-        ]
-        no_watch_queries = [
-            group
-            for group in enabled_groups
-            if group.is_tag_group and not group.queries
-        ]
-        no_push_targets = [group for group in enabled_groups if not group.targets]
-        no_check_triggers = [
-            group
-            for group in enabled_groups
-            if not group.interval_check_enabled and not group.daily_check_enabled
-        ]
-
-        if no_watch_users:
-            items.append(
-                {
-                    "key": "groups_without_watch_users",
-                    "level": "warning",
-                    "title": "启用博主分组没有关注账号",
-                    "detail": (
-                        "这些分组不会检查任何账号："
-                        + NitterWebAPI._format_group_names(no_watch_users)
-                    ),
-                }
-            )
-        if no_watch_queries:
-            items.append(
-                {
-                    "key": "groups_without_watch_queries",
-                    "level": "warning",
-                    "title": "启用标签分组没有搜索订阅",
-                    "detail": (
-                        "这些分组不会检查任何查询："
-                        + NitterWebAPI._format_group_names(no_watch_queries)
-                    ),
-                }
-            )
-        if no_push_targets:
-            items.append(
-                {
-                    "key": "groups_without_push_targets",
-                    "level": "warning",
-                    "title": "启用分组没有推送目标",
-                    "detail": (
-                        "新推文无法送达这些分组："
-                        + NitterWebAPI._format_group_names(no_push_targets)
-                    ),
-                }
-            )
-        if no_check_triggers:
-            items.append(
-                {
-                    "key": "groups_without_check_triggers",
-                    "level": "warning",
-                    "title": "启用分组没有检查触发",
-                    "detail": (
-                        "这些分组既没有间隔检查也没有每日检查："
-                        + NitterWebAPI._format_group_names(no_check_triggers)
-                    ),
-                }
-            )
-        return items
-
-    @staticmethod
-    def _format_group_names(groups: list[ScheduleGroup], limit: int = 4) -> str:
-        names = [str(group.name or group.group_id) for group in groups]
-        visible = names[:limit]
-        suffix = f" 等 {len(names)} 个" if len(names) > limit else ""
-        return "、".join(visible) + suffix
-
-    def _serialize_group(
-        self,
-        group: ScheduleGroup,
-    ) -> dict[str, Any]:
-        return {
-            "group_id": group.group_id,
-            "name": group.name,
-            "enabled": group.enabled,
-            "group_type": group.group_type,
-            "aliases": list(group.aliases),
-            "watch_users": list(group.users),
-            "watch_user_count": len(group.users),
-            "raw_watch_user_count": group.users_info.raw_count,
-            "duplicate_watch_users": list(group.users_info.duplicates),
-            "invalid_watch_users": list(group.users_info.invalid_entries),
-            "watch_queries": [
-                {"query": item.query, "type": item.type} for item in group.queries
-            ],
-            "watch_query_count": len(group.queries),
-            "raw_watch_query_count": group.queries_info.raw_count,
-            "duplicate_watch_queries": list(group.queries_info.duplicates),
-            "invalid_watch_queries": list(group.queries_info.invalid_entries),
-            "push_targets": list(group.targets),
-            "push_target_count": len(group.targets),
-            "invalid_push_targets": list(group.invalid_targets),
-            "invalid_push_target_count": len(group.invalid_targets),
-            "interval_check_enabled": group.interval_check_enabled,
-            "check_interval_minutes": group.check_interval_minutes,
-            "daily_check_enabled": group.daily_check_enabled,
-            "daily_check_times": self._format_times(group.daily_check_times),
-            "filter_plain_text_enabled": group.filter_plain_text_enabled,
-            "media_only_enabled": group.media_only_enabled,
-            "omit_status_url": bool(getattr(group, "omit_status_url", True)),
-            "hide_original_when_translated": bool(
-                getattr(group, "hide_original_when_translated", False)
-            ),
-            "media_only_effective": self._media_only_effective(group),
-            # Global media availability only; independent of saved group toggle
-            # so the dashboard draft can warn before save.
-            "media_only_unavailable_reason": media_only_unavailable_reason(self.config),
-            "attention_items": self._group_attention_items(group),
-        }
-
-    def _media_only_effective(self, group: ScheduleGroup) -> bool:
-        return bool(
-            group.media_only_enabled and not media_only_unavailable_reason(self.config)
-        )
-
-    @staticmethod
-    def _group_attention_items(
-        group: ScheduleGroup,
-    ) -> list[dict[str, str]]:
-        items: list[dict[str, str]] = []
-        if not group.enabled:
-            items.append(
-                {
-                    "key": "group_disabled",
-                    "level": "info",
-                    "title": "分组停用",
-                    "detail": "停用分组不会参与后台检查。",
-                }
-            )
-        if group.is_tag_group:
-            if not group.queries:
-                items.append(
-                    {
-                        "key": "no_watch_queries",
-                        "level": "warning",
-                        "title": "无搜索订阅",
-                        "detail": "该标签分组没有可检查的搜索订阅。",
-                    }
-                )
-        elif not group.users:
-            items.append(
-                {
-                    "key": "no_watch_users",
-                    "level": "warning",
-                    "title": "无关注账号",
-                    "detail": "该分组没有可检查的关注账号。",
-                }
-            )
-        if not group.targets:
-            items.append(
-                {
-                    "key": "no_push_targets",
-                    "level": "warning",
-                    "title": "无推送目标",
-                    "detail": "新推文没有可发送的推送目标。",
-                }
-            )
-        if group.users_info.invalid_entries:
-            items.append(
-                {
-                    "key": "invalid_watch_users",
-                    "level": "warning",
-                    "title": "关注账号无效",
-                    "detail": f"{len(group.users_info.invalid_entries)} 个关注账号格式无效。",
-                }
-            )
-        if group.invalid_targets:
-            items.append(
-                {
-                    "key": "invalid_push_targets",
-                    "level": "warning",
-                    "title": "推送目标无效",
-                    "detail": f"{len(group.invalid_targets)} 个推送目标未通过 UMO 校验。",
-                }
-            )
-        return items
-
-    @staticmethod
-    def _serialize_history_record(
-        record: PushHistoryRecord,
-        group_names: dict[str, str],
-    ) -> dict[str, Any]:
-        tweet = record.tweet
-        return {
-            "id": record.id,
-            "group_id": record.group_id,
-            "group_name": group_names.get(record.group_id, record.group_id),
-            "username": record.username,
-            "status_id": record.status_id,
-            "original_link": record.original_link or tweet.x_url,
-            "target_umo": record.target_umo,
-            "source": record.source,
-            "instance": record.instance,
-            "pushed_at": record.pushed_at,
-            "delivery_status": record.delivery_status,
-            "delivery_error": record.delivery_error,
-            "published": tweet.published,
-            "text_preview": NitterWebAPI._text_preview(tweet.text),
-            "translation_preview": NitterWebAPI._text_preview(tweet.translation),
-        }
-
-    @staticmethod
-    def _serialize_history_group_summary(
-        summary: PushHistoryGroupSummary,
-    ) -> dict[str, Any]:
-        return {
-            "group_id": summary.group_id,
-            "record_count": summary.record_count,
-            "user_count": summary.user_count,
-            "latest_pushed_at": summary.latest_pushed_at,
-        }
-
-    @staticmethod
-    def _group_history_records(
-        records: list[PushHistoryRecord],
-        group_names: dict[str, str],
-        groups_by_id: dict[str, ScheduleGroup],
-    ) -> list[dict[str, Any]]:
-        grouped: dict[tuple[str, str, str, str, str], dict[str, Any]] = {}
-        order: list[tuple[str, str, str, str, str]] = []
-        latest_delivery_by_key: dict[
-            tuple[str, str, str, str, str],
-            dict[str, tuple[int, int, str, str]],
-        ] = {}
-        for record in records:
-            key = (
-                record.group_id,
-                record.username,
-                record.status_id,
-                record.source,
-                record.original_link or record.tweet.x_url,
-            )
-            item = grouped.get(key)
-            serialized = NitterWebAPI._serialize_history_record(record, group_names)
-            if item is None:
-                group = groups_by_id.get(record.group_id)
-                current_targets = list(getattr(group, "targets", []) or [])
-                group_type = str(
-                    getattr(group, "group_type", "") or ""
-                ).strip().lower()
-                if group_type not in {"blogger", "tag"}:
-                    group_type = (
-                        "tag"
-                        if str(record.username or "").strip().lower().startswith("q:")
-                        else "blogger"
-                    )
-                item = {
-                    **serialized,
-                    "group_type": group_type,
-                    "target_umos": [],
-                    "target_count": 0,
-                    "replay_target_options": [
-                        {
-                            "umo": target,
-                            "historical": False,
-                            "available": True,
-                        }
-                        for target in current_targets
-                    ],
-                }
-                grouped[key] = item
-                order.append(key)
-                latest_delivery_by_key[key] = {}
-            if record.target_umo and record.target_umo not in item["target_umos"]:
-                item["target_umos"].append(record.target_umo)
-            item["target_count"] = len(item["target_umos"])
-            if int(record.pushed_at or 0) > int(item.get("pushed_at") or 0):
-                item.update(serialized)
-            target_key = record.target_umo or f"__record__:{record.id}"
-            delivery_stamp = (int(record.pushed_at or 0), int(record.id or 0))
-            latest_delivery = latest_delivery_by_key[key]
-            previous = latest_delivery.get(target_key)
-            if previous is None or delivery_stamp > previous[:2]:
-                latest_delivery[target_key] = (
-                    delivery_stamp[0],
-                    delivery_stamp[1],
-                    record.delivery_status,
-                    record.delivery_error,
-                )
-            options_by_umo = {
-                option["umo"]: option for option in item["replay_target_options"]
-            }
-            if record.target_umo:
-                option = options_by_umo.get(record.target_umo)
-                if option is None:
-                    item["replay_target_options"].append(
-                        {
-                            "umo": record.target_umo,
-                            "historical": True,
-                            "available": False,
-                        }
-                    )
-                else:
-                    option["historical"] = True
-        for key, item in grouped.items():
-            latest_states = list(latest_delivery_by_key[key].values())
-            has_partial = any(
-                status == "partial_failed"
-                for _pushed_at, _record_id, status, _error in latest_states
-            )
-            partial_errors = list(
-                dict.fromkeys(
-                    str(error).strip()
-                    for _pushed_at, _record_id, status, error in latest_states
-                    if status == "partial_failed" and str(error).strip()
-                )
-            )
-            if has_partial:
-                item["delivery_status"] = "partial_failed"
-                item["delivery_error"] = "; ".join(partial_errors)
-            else:
-                item["delivery_status"] = "success"
-                item["delivery_error"] = ""
-        return [grouped[key] for key in order]
-
-    @staticmethod
-    def _serialize_check_result(result: Any) -> dict[str, Any]:
-        return {
-            "group_id": getattr(result, "group_id", ""),
-            "group_name": getattr(result, "group_name", ""),
-            "skipped_reason": getattr(result, "skipped_reason", ""),
-            "new_tweet_count": getattr(result, "new_tweet_count", 0),
-            "pushed_target_successes": getattr(result, "pushed_target_successes", 0),
-            "pushed_target_attempts": getattr(result, "pushed_target_attempts", 0),
-        }
-
-    @staticmethod
-    def _serialize_cache_result(result: Any) -> dict[str, int]:
-        return {
-            "removed": int(getattr(result, "removed", 0) or 0),
-            "failed": int(getattr(result, "failed", 0) or 0),
-            "skipped_dirs": int(getattr(result, "skipped_dirs", 0) or 0),
-            "removed_images": int(
-                getattr(result, "removed_images", getattr(result, "images", 0)) or 0
-            ),
-            "removed_videos": int(
-                getattr(result, "removed_videos", getattr(result, "videos", 0)) or 0
-            ),
-            "removed_other": int(
-                getattr(result, "removed_other", getattr(result, "other", 0)) or 0
-            ),
-            "skipped_active": int(getattr(result, "skipped_active", 0) or 0),
-            "removed_empty_dirs": int(getattr(result, "removed_empty_dirs", 0) or 0),
-        }
-
-    @staticmethod
-    def _serialize_probe_tweet(tweet: TweetItem) -> dict[str, Any]:
-        return {
-            "status_id": tweet.status_id,
-            "username": tweet.username,
-            "link": tweet.x_url,
-            "published": tweet.published,
-            "text_preview": NitterWebAPI._text_preview(tweet.text),
-            "media_count": len(tweet.media),
-        }
-
-    @staticmethod
-    def _format_times(times: list[tuple[int, int]]) -> list[str]:
-        return [f"{hour:02d}:{minute:02d}" for hour, minute in times]
-
-    @staticmethod
-    def _text_preview(text: str, limit: int = 160) -> str:
-        text = " ".join(str(text or "").split())
-        return text if len(text) <= limit else text[: limit - 1] + "…"
-
-    @staticmethod
-    def _terminology() -> dict[str, str]:
-        return {
-            "watch_users": "关注账号",
-            "push_targets": "推送目标",
-            "seen": "推送记录",
-        }
-
-    @staticmethod
-    def _group_label(group: ScheduleGroup | None) -> str:
-        if group is None:
-            return "全部分组"
-        return f"{group.name} ({group.group_id})"
-
-    @staticmethod
-    def _group_payload_label(group: ScheduleGroup) -> dict[str, str]:
-        return {"group_id": group.group_id, "name": group.name}
-
-    @staticmethod
-    def _data_text(data: dict[str, Any], key: str) -> str:
-        return str(data.get(key, "") or "").strip()
-
-    @staticmethod
-    def _data_text_list(data: dict[str, Any], key: str) -> list[str]:
-        value = data.get(key, [])
-        if isinstance(value, str):
-            return [item.strip() for item in value.split(",") if item.strip()]
-        if isinstance(value, list):
-            return [str(item).strip() for item in value if str(item).strip()]
-        return []
-
-    @staticmethod
-    def _parse_int(
-        value: Any,
-        fallback: int,
-        *,
-        minimum: int,
-        maximum: int,
-    ) -> int:
-        try:
-            number = int(value)
-        except (TypeError, ValueError):
-            number = fallback
-        return max(minimum, min(maximum, number))
-
-    @staticmethod
-    def _parse_float(
-        value: Any,
-        fallback: float,
-        *,
-        minimum: float,
-        maximum: float,
-    ) -> float:
-        try:
-            number = float(value)
-        except (TypeError, ValueError):
-            number = fallback
-        return max(minimum, min(maximum, number))
 
     @staticmethod
     def _ok(**payload) -> dict[str, Any]:
